refactor ♻️: Replace CustomResponseMethodsTrait with HasCommonApiResponses and InteractsWithResponseMethods traits
- Removed CustomResponseMethodsTrait to consolidate response handling into HasCommonApiResponses and InteractsWithResponseMethods traits.
- Introduced new methods for standardized API responses, improving code organization and reusability.
- Enhanced metadata generation with a dedicated method in InteractsWithResponseMethods, streamlining response structure.
